refactor(shared): share changed-file parsing across activity consumers
